backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python>=4.8.0
pillow>=10.0.0
easyocr>=1.7.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0,<2.0.0
tensorflow==2.15.0
ultralytics>=8.0.0
scikit-learn>=1.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
pydantic==2.5.0